Remove dockerhub login to stop pushing to it
Remove dockerhub login as its uneeded if only want to push to ghcr.io.
It's been pushing to both.

Will also help me and other forking by not having to edit and remove docker hub to run the action, as I don't want to use it or have to setup each time I fork.